+++ internal/module/credit_limit/entity/entity.go
@@ -1,7 +1,7 @@
 package entity
 
 type CreditLimit struct {
-	CustomerID  int     `db:"customer_id"`
+	CustomerID  int64   `db:"customer_id"`
 	TenorMonth  int     `db:"tenor_month"`
 	LimitAmount float64 `db:"limit_amount"`
 }
