+++ pkg/jwt_handler/types.go
@@ -3,7 +3,7 @@ package jwt_handler
 import "github.com/golang-jwt/jwt/v5"
 
 type CustomClaims struct {
-	CustomerID int    `json:"customer_id"`
+	CustomerID int64  `json:"customer_id"`
 	Nik        string `json:"nik"`
 	Email      string `json:"email"`
 	FullName   string `json:"full_name"`
@@ -11,7 +11,7 @@ type CustomClaims struct {
 }
 
 type CostumClaimsPayload struct {
-	CustomerID int    `json:"customer_id"`
+	CustomerID int64  `json:"customer_id"`
 	Nik        string `json:"nik"`
 	Email      string `json:"email"`
 	FullName   string `json:"full_name"`
